refactor(domain): 金额全链路 BigDecimal
BillFilters/CsvImport/仓储/异常检测的金额语义从 double 换为
BigDecimal(输入-存储-计算一致,HALF_EVEN);对外 JSON 数字形状不变。

+++ billguard-domain/src/main/java/io/github/lxyang01/billguard/bills/BillFilters.java
@@ -2,6 +2,7 @@ package io.github.lxyang01.billguard.bills;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
+import java.math.BigDecimal;
 import java.util.Set;
 
 /** 交易筛选条件(字段与 BillFilters 一致)。 */
@@ -12,8 +13,8 @@ public record BillFilters(
     String merchant,
     String method,
     String status,
-    Double minAmount,
-    Double maxAmount,
+    BigDecimal minAmount,
+    BigDecimal maxAmount,
     String query) {
 
     public static final BillFilters EMPTY = new BillFilters("", "", "", "", "", "", null,
@@ -37,8 +38,8 @@ public record BillFilters(
             text(picked.get("date_from")), text(picked.get("date_to")),
             text(picked.get("category")), text(picked.get("merchant")),
             text(picked.get("method")), text(picked.get("status")),
-            picked.get("min_amount") instanceof Number n ? n.doubleValue() : null,
-            picked.get("max_amount") instanceof Number n ? n.doubleValue() : null,
+            picked.get("min_amount") instanceof Number n ? BigDecimal.valueOf(n.doubleValue()) : null,
+            picked.get("max_amount") instanceof Number n ? BigDecimal.valueOf(n.doubleValue()) : null,
             text(picked.get("query")));
     }
